requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15,<4.0
django-environ>=0.11.2
python-decouple==3.8
Pillow>=10.3.0
psycopg2-binary>=2.9.9
celery==5.3.4
redis==5.0.1
pytest==7.4.3
pytest-django>=4.8
pytest-cov==4.1.0
factory-boy==3.3.0
faker==21.0.0
WeasyPrint>=63.1
django-filter>=24.0
django-cors-headers==4.3.1
python-dateutil==2.8.2
pytz==2023.3
black==23.12.1
ruff==0.1.9
